Dodani testovi za UserDelete
